requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
distlib==0.3.7
distro==1.8.0
Django==4.1.6
django-cors-headers==4.2.0
django-extra-fields==3.0.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.12.3
flatbuffers==23.3.3
frozenlist==1.3.3
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
gast==0.4.0
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.6.0
incremental==22.10.0
ipinfo==4.4.2
jax==0.4.8
Keras==2.4.3
libclang==16.0.0
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
multidict==6.0.4
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
paho-mqtt==1.6.1
Pillow==9.4.0
pipenv==2023.9.1
platformdirs==3.10.0
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygit2==1.12.2
pyOpenSSL==23.1.1
pyTelegramBotAPI==4.12.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
service-identity==21.1.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.3
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-macos==2.12.0
termcolor==2.2.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2022.7
ujson==5.7.0
urllib3==1.26.14
virtualenv==20.24.4
Werkzeug==2.2.3
whitenoise==6.3.0
wrapt==1.14.1
yarl==1.9.2
zipp==3.15.0
zope.interface==6.0
